protecting keys
