+++ config/connection.js
@@ -3,9 +3,9 @@ var mysql = require("mysql");
 
 var connection = mysql.createConnection({
   host: "127.0.0.1",
-  port: 3307,
+  port: 3306,
   user: "root",
-  password: "root",
+  password: "",
   database: "saved_receipesDB"
 });
