+++ server.js
@@ -1,3 +1,4 @@
+require("dotenv").config();
 var express = require("express");
 
 var PORT = process.env.PORT || 3000;
